+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.121.0
 
 - `[initsystem/sdnotify]` Added new package for sending messages to systemd
+- `[support/deps]` Updated for compatibility with the latest version of [depsy](https://kaos.sh/depsy)
 
 ### 12.120.0